streaming: split streaming_test.go into per-concern test files
Pure test reorganization, no logic change — the 894-line omnibus split by
concern, function bodies moved verbatim:

- window_test.go            geometry / window arithmetic
- keys_test.go              key schema + key<->path bijection
- catalog_test.go           catalog reader (pins, scans, FrozenCoverage)
- catalog_protocol_test.go  mark/flip + CommitIndex
- catalog_sweep_test.go     the two sweeps
- crashsafety_test.go       power-loss-between-steps + never-unlink-under-frozen
- helpers_test.go           shared test helpers + vars

Same 33 tests (51 package-wide); gofmt clean, vet/test green.
